refactor(PrizeLevel): 패키지 변경
프로그램의 root에는 Application과 LottoGame만 있도록 변경
