+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,11 +1,10 @@
-package lotto;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.constants.LottoConstants;
 import lotto.constants.LottoMachineConstants;
-import lotto.domain.Lotto;
 
 public class LottoMachine {
     private static final int LOTTO_PRICE = LottoMachineConstants.LOTTO_PRICE;
@@ -36,7 +35,7 @@ public class LottoMachine {
         System.out.println();
         System.out.println(lottos.size() + "개를 구매했습니다.");
         for (Lotto lotto : lottos) {
-            System.out.println(lotto);
+            System.out.println(lotto.getNumbers());
         }
     }
 }
